+++ src/onegov.ballot/onegov/ballot/tests/models/test_vote.py
@@ -404,36 +404,74 @@ def test_vote_turnout(session):
     assert vote.proposal.turnout == 10
 
 
-def test_vote_percentage_by_entity(session):
+def test_vote_results_by_district(session):
     vote = Vote(
         title="Abstimmung",
         domain='federation',
         date=date(2015, 6, 18)
     )
-
     vote.ballots.append(Ballot(type='proposal'))
-
     session.add(vote)
     session.flush()
+    assert vote.proposal.results_by_district.all() == []
 
     vote.proposal.results.append(
         BallotResult(
             name='1', entity_id=1,
-            counted=True, eligible_voters=100, yeas=75, nays=25
+            counted=True, eligible_voters=100,
+            yeas=75, nays=25
         )
     )
-    vote.proposal.results.append(
-        BallotResult(
-            name='1', entity_id=1,
-            counted=True, eligible_voters=100, yeas=25, nays=75
-        )
-    )
-
     session.flush()
+    assert [list(r) for r in vote.proposal.results_by_district] == [
+        [None, True, True, 75, 25, 75.0, 25.0, 0, 0, 100]
+    ]
 
-    assert vote.proposal.percentage_by_entity() == {
-        1: {'counted': True, 'yeas_percentage': 50.0, 'nays_percentage': 50.0}
-    }
+    vote.proposal.results.append(
+        BallotResult(
+            name='2', entity_id=2,
+            counted=False, eligible_voters=100,
+            yeas=25, nays=75
+        )
+    )
+    session.flush()
+    assert [list(r) for r in vote.proposal.results_by_district] == [
+        [None, False, None, 100, 100, 50.0, 50.0, 0, 0, 200]
+    ]
+
+    vote.proposal.results.append(
+        BallotResult(
+            name='1', entity_id=1, district='a',
+            counted=True, eligible_voters=100,
+            empty=1, invalid=2, yeas=10, nays=30
+        )
+    )
+    vote.proposal.results.append(
+        BallotResult(
+            name='2', entity_id=2, district='a',
+            counted=True, eligible_voters=200,
+            empty=3, invalid=4, yeas=50, nays=10
+        )
+    )
+    vote.proposal.results.append(
+        BallotResult(
+            name='3', entity_id=3, district='b',
+            counted=True, eligible_voters=300,
+            empty=5, invalid=6, yeas=30, nays=10
+        )
+    )
+    vote.proposal.results.append(
+        BallotResult(
+            name='4', entity_id=4, district='b',
+            counted=True, eligible_voters=400,
+            empty=7, invalid=8, yeas=10, nays=50
+        )
+    )
+    assert [list(r) for r in vote.proposal.results_by_district] == [
+        ['a', True, True, 60, 40, 60.0, 40.0, 4, 6, 300],
+        ['b', True, False, 40, 60, 40.0, 60.0, 12, 14, 700],
+        [None, False, None, 100, 100, 50.0, 50.0, 0, 0, 200]
+    ]
 
 
 def test_ballot_results_aggregation(session):